// src/apis/dataType.ts
export interface ProjectType {
    p_no: number
    p_title: string
    p_desc: string
    p_date: string
    p_person: string
    p_party: string
    p_language: string
    p_url?: string
    p_img1: string | null
    p_img2: string | null
    p_img3: string | null
    p_img4: string | null
    p_vide1?: string | null
    p_vide2?: string | null
    p_vide3?: string | null
    p_git: string
}